+++ src/lib/pages/contract-details/index.tsx
@@ -5,16 +5,14 @@ import {
   Tabs,
   TabPanels,
   TabPanel,
-  Icon,
-  Text,
 } from "@chakra-ui/react";
 import { observer } from "mobx-react-lite";
 import { useRouter } from "next/router";
-import { MdSearchOff } from "react-icons/md";
 
 import { BackButton } from "lib/components/button/BackButton";
 import { CustomTab } from "lib/components/CustomTab";
 import PageContainer from "lib/components/PageContainer";
+import { InvalidState } from "lib/components/state/InvalidState";
 import { useValidateAddress } from "lib/hooks";
 import {
   useContractData,
@@ -36,26 +34,7 @@ interface ContractDetailsBodyProps {
   contractAddress: ContractAddr;
 }
 
-const InvalidContract = () => (
-  <Flex
-    direction="column"
-    alignItems="center"
-    borderY="1px solid"
-    borderColor="divider.main"
-    width="full"
-    my="24px"
-    py="24px"
-  >
-    <Icon as={MdSearchOff} color="gray.600" boxSize="128" />
-    <Heading as="h5" variant="h5" my="8px">
-      Contract does not exist
-    </Heading>
-    <Text variant="body2" fontWeight="500" color="gray.500" textAlign="center">
-      Please double-check your spelling and make sure you have selected the
-      correct network.
-    </Text>
-  </Flex>
-);
+const InvalidContract = () => <InvalidState title="Contract does not exist" />;
 
 const ContractDetailsBody = observer(
   ({ contractAddress }: ContractDetailsBodyProps) => {