+++ src/lib/services/codeService.ts
@@ -12,7 +12,7 @@ import {
 } from "lib/data/queries";
 import type {
   CodeInfo,
-  CodeDetails,
+  CodeData,
   ContractAddr,
   Option,
   ContractInfo,
@@ -81,7 +81,7 @@ export const useCodeListByIDsQuery = (ids: Option<number[]>) => {
 
 export const useCodeInfoByCodeId = (
   codeId: Option<number>
-): UseQueryResult<Option<Omit<CodeDetails, "chainId">>> => {
+): UseQueryResult<Option<Omit<CodeData, "chainId">>> => {
   const queryFn = useCallback(async () => {
     if (!codeId) return undefined;